requirements.txt
addict==2.4.0
ansible-cmdb==1.31
argparse==1.4.0
attrs==22.1.0
autocommand==2.2.2
batchgenerators==0.23
brotlipy==0.7.0
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.4
cheroot==9.0.0
cherrypy==18.8.0
click==8.1.3
cloudpickle==2.2.0
cryptography==36.0.0
cycler==0.11.0
docker-pycreds==0.4.0
einops==0.4.1
fonttools==4.33.3
future==0.18.2
gitdb==4.0.9
gitpython==3.1.27
h5py==3.6.0
hostlist==1.4.8
hyperopt==0.2.7
idna==3.3
imageio==2.19.0
inflect==6.0.2
jaraco.classes==3.2.3
jaraco.collections==3.8.0
jaraco.context==4.2.0
jaraco.functools==3.5.2
jaraco.text==3.11.0
joblib==1.1.0
jsonref==0.2
jsonschema==4.16.0
jsonxs==0.6
kiwisolver==1.4.2
linecache2==1.0.0
mako==1.2.4
markupsafe==2.1.1
matplotlib==3.5.2
medpy==0.4.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mmcv==1.5.0
git+https://github.com/Project-MONAI/MONAI.git@07de215c
more-itertools==9.0.0
networkx==2.8
nibabel==3.2.2
numpy==1.21.5
opencv-python==4.5.5.64
packaging==21.3
panda==0.3.1
pandas==1.4.2
pathtools==0.1.2
pillow==9.0.1
pip==21.2.4
portend==3.1.0
promise==2.3
protobuf==3.20.0rc2
psutil==5.9.1
py4j==0.10.9.7
pycparser==2.21
pydantic==1.10.2
pydicom==2.3.0
pyopenssl==22.0.0
pyparsing==3.0.8
pyrsistent==0.18.1
pysocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
pywavelets==1.3.0
pyyaml==6.0
requests==2.27.1
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
sentry-sdk==1.5.12
setproctitle==1.2.3
setuptools==61.2.0
shortuuid==1.0.9
simpleitk==2.1.1.2
six==1.16.0
smmap==5.0.0
sweeps==0.2.0
tempora==5.1.0
tensorboardx==2.5
termcolor==2.1.1
threadpoolctl==3.1.0
tifffile==2022.5.4
timm==0.4.12
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.64.0
traceback2==1.4.0
typing-extensions==4.1.1
typing==3.7.4.3
unicode==2.9
unidecode==1.3.4
unittest2==1.1.0
urllib3==1.26.9
ushlex==0.99.1
wandb==0.13.3
wheel==0.37.1
yapf==0.32.0
zc.lockfile==2.0